+++ tests/testthat/test-json_design.R
@@ -4,14 +4,14 @@ context("test-json_design")
 test_that("defaults", {
   des <- check_design(mu = 1.123456789, plot = FALSE)
   json <- json_design(des)
-  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.12345679}},"sd":{"y":{"y":1}},"r":[]}'
+  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.12345679}},"sd":{"y":{"y":1}},"r":[],"sep":"_"}'
   class(txt) <- "json"
   
   expect_equal(json, txt)
   
   des <- check_design(2,2, plot = FALSE)
   json <- json_design(des)
-  txt <- '{"within":{"A":{"A1":"A1","A2":"A2"}},"between":{"B":{"B1":"B1","B2":"B2"}},"dv":{"y":"value"},"id":{"id":"id"},"n":{"B1":100,"B2":100},"mu":{"B1":{"A1":0,"A2":0},"B2":{"A1":0,"A2":0}},"sd":{"B1":{"A1":1,"A2":1},"B2":{"A1":1,"A2":1}},"r":{"B1":[[1,0],[0,1]],"B2":[[1,0],[0,1]]}}'
+  txt <- '{"within":{"A":{"A1":"A1","A2":"A2"}},"between":{"B":{"B1":"B1","B2":"B2"}},"dv":{"y":"value"},"id":{"id":"id"},"n":{"B1":100,"B2":100},"mu":{"B1":{"A1":0,"A2":0},"B2":{"A1":0,"A2":0}},"sd":{"B1":{"A1":1,"A2":1},"B2":{"A1":1,"A2":1}},"r":{"B1":[[1,0],[0,1]],"B2":[[1,0],[0,1]]},"sep":"_"}'
   class(txt) <- "json"
   
   expect_equal(json, txt)
@@ -44,13 +44,13 @@ test_that("filename", {
 test_that("digits", {
   des <- check_design(mu = 1.123456789, plot = FALSE)
   json <- json_design(des, digits = 3)
-  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.123}},"sd":{"y":{"y":1}},"r":[]}'
+  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.123}},"sd":{"y":{"y":1}},"r":[],"sep":"_"}'
   class(txt) <- "json"
   
   expect_equal(json, txt)
   
   json <- json_design(des, digits = 4)
-  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.1235}},"sd":{"y":{"y":1}},"r":[]}'
+  txt <- '{"within":[],"between":[],"dv":{"y":"value"},"id":{"id":"id"},"n":{"y":100},"mu":{"y":{"y":1.1235}},"sd":{"y":{"y":1}},"r":[],"sep":"_"}'
   class(txt) <- "json"
   
   expect_equal(json, txt)
@@ -82,7 +82,8 @@ test_that("pretty", {
       "y": 1
     }
   },
-  "r": []
+  "r": [],
+  "sep": "_"
 }'
   
   class(txt) <- "json"